Re-factored to use SidedProxy model
