+++ src/main/java/lakmoore/infinitic/lib/data/FluidDataJSON.java
@@ -1,7 +1,7 @@
-package me.johz.infinitic.lib.data;
+package lakmoore.infinitic.lib.data;
 
-import me.johz.infinitic.lib.errors.JSONValidationException;
-import me.johz.infinitic.lib.helpers.GenericHelper;
+import lakmoore.infinitic.lib.errors.JSONValidationException;
+import lakmoore.infinitic.lib.helpers.GenericHelper;
 import net.minecraft.item.EnumRarity;
 
 public class FluidDataJSON implements IJson {
