+++ src/main/java/lakmoore/infinitic/lib/data/MaterialData.java
@@ -1,4 +1,4 @@
-package me.johz.infinitic.lib.data;
+package lakmoore.infinitic.lib.data;
 
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 import net.minecraftforge.fml.relauncher.Side;
@@ -6,10 +6,9 @@ import net.minecraftforge.oredict.OreDictionary;
 
 import java.util.ArrayList;
 
-import me.johz.infinitic.InfiniTiC;
-import me.johz.infinitic.client.model.InfiniFluidStateMapper;
-import me.johz.infinitic.lib.errors.JSONValidationException;
-import me.johz.infinitic.lib.helpers.GenericHelper;
+import lakmoore.infinitic.InfiniTiC;
+import lakmoore.infinitic.lib.errors.JSONValidationException;
+import lakmoore.infinitic.lib.helpers.GenericHelper;
 import net.minecraft.block.Block;
 import slimeknights.tconstruct.library.materials.ArrowShaftMaterialStats;
 import slimeknights.tconstruct.library.materials.BowMaterialStats;
@@ -27,12 +26,10 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.fluids.Fluid;
 import net.minecraftforge.fluids.FluidRegistry;
 import slimeknights.tconstruct.library.MaterialIntegration;
 import slimeknights.tconstruct.library.TinkerRegistry;
-import slimeknights.tconstruct.library.client.MaterialRenderInfo;
 import slimeknights.tconstruct.library.fluid.FluidMolten;
 import slimeknights.tconstruct.smeltery.TinkerSmeltery;
 import slimeknights.tconstruct.smeltery.block.BlockMolten;
@@ -72,38 +69,8 @@ public class MaterialData {
 		}
 	}
 	
-	public void preInit(Side side) {
-	    if (isValid)
-	    {
-	    		getSolids();
-	    		addToOreDict();
-	    		makeMaterial();
-	    		if(json.getMake("fluid")) {
-	    	        makeFluid();	    			
-	    		}
-	    		integrateMaterial();
-
-	    		if(side == Side.CLIENT) {
-	    			
-	    			if (fluid != null) {
-		    		    Item fluidItem = Item.getItemFromBlock(block);
-		    		    InfiniFluidStateMapper mapper = new InfiniFluidStateMapper(fluid);
-		    		    // item-model
-		    		    ModelLoader.registerItemVariants(fluidItem);
-		    		    ModelLoader.setCustomMeshDefinition(fluidItem, mapper);
-		    		    // block-model
-		    		    ModelLoader.setCustomStateMapper(block, mapper);	    				
-	    			}
-
-			    int color = json.getToolColorInt();
-			    float shinyness = 0.25f;
-			    float brightness = 0.5f;
-			    float hueshift = -0.1f;	    
-			    material.setRenderInfo(new MaterialRenderInfo.Metal(color, shinyness, brightness, hueshift));
-			}
-
-	    		addMaterialStats();
-	    }
+	public boolean isValid() {
+		return isValid;
 	}
 
 	public void init(Side side) {
@@ -117,15 +84,8 @@ public class MaterialData {
 			}
 	    }		
 	}
-	
-	public void postInit(Side side) {
-	    if (isValid)
-	    {
-		    if(integration != null) integration.integrate();	    	
-	    }
-	}
 
-	private void addToOreDict() {
+	public void addToOreDict() {
 		String suffix = GenericHelper.capitalizeFirstLetter(json.name);
 		String[] types = {"ingot", "nugget", "gem", "dust", "block", "ore"};
 		
@@ -136,7 +96,7 @@ public class MaterialData {
 		}
 	}
 
-	private void integrateMaterial() {
+	public void integrateMaterial() {
 		
 		//if there is no fluid then we can only use the part builder
 		if (fluid == null) {
@@ -167,7 +127,7 @@ public class MaterialData {
 		integration.preInit();
 	}
 
-	private void makeMaterial() {
+	public void makeMaterial() {
 		
 		material = TinkerRegistry.getMaterial(json.name);
 		if (material != Material.UNKNOWN) {
@@ -180,7 +140,7 @@ public class MaterialData {
 
 	}
 
-	private void getSolids()
+	public void getSolids()
 	{
 		//Get the ore block		
 		//TODO: If no ore block exists, make one???
@@ -199,7 +159,7 @@ public class MaterialData {
 
 	}
 	
-	private void addTraitType(String[] traitsList, String dependency) {
+	public void addTraitType(String[] traitsList, String dependency) {
 		ArrayList<String> traitsSoFar = new ArrayList<String>();
 
 		for (String traitId : traitsList) {
@@ -219,7 +179,7 @@ public class MaterialData {
 		}
 	}
 
-	private void addMaterialTraits() {			
+	public void addMaterialTraits() {			
 		addTraitType(json.toolData.headTraits, MaterialTypes.HEAD);
 		addTraitType(json.toolData.handleTraits, MaterialTypes.HANDLE);
 		addTraitType(json.toolData.extraTraits, MaterialTypes.EXTRA);
@@ -231,17 +191,17 @@ public class MaterialData {
 		addTraitType(json.toolData.traits, "");
 	}
 	
-	private void addMaterialStats() {
+	public void addMaterialStats() {
 		
 		//can we make heads out of this material?
-	    if(json.toolData.miningSpeed != 0 || json.toolData.attack != 0) {
+	    if(json.toolData.miningSpeed != 0 && json.toolData.attack != 0 && json.toolData.durability != 0 && json.toolData.harvestLevel != 0) {
 		    TinkerRegistry.addMaterialStats(material,
 	                new HeadMaterialStats(json.toolData.durability, json.toolData.miningSpeed, json.toolData.attack, json.toolData.harvestLevel)
 	                );
 	    }
 
 		//can we make handles out of this material?
-	    if(json.toolData.handleModifier != 0) {
+	    if(json.toolData.handleModifier != 0 && json.toolData.handleDurability != 0) {
 		    TinkerRegistry.addMaterialStats(material,
 	                new HandleMaterialStats(json.toolData.handleModifier, json.toolData.handleDurability)
 	                );
@@ -269,7 +229,7 @@ public class MaterialData {
 	    }
 
 	    //can we make arrow fletchings out of this material?
-	    if (json.toolData.accuracy != 0 || json.toolData.fletchingModifier != 0) {
+	    if (json.toolData.accuracy != 0 && json.toolData.fletchingModifier != 0) {
 		    TinkerRegistry.addMaterialStats(material,
 			    	new FletchingMaterialStats(json.toolData.accuracy, json.toolData.fletchingModifier)
 			    	);
@@ -285,7 +245,7 @@ public class MaterialData {
 	    //can we make projectiles (e.g. Shurikens) out of this material?
 	    //Tinkers auto-adds this stat to any material used to make tool heads
 	    //and trying to add it a second time throws an exception, so check before adding.
-	    if (json.toolData.projectiles && !material.hasStats(MaterialTypes.PROJECTILE)) {
+	    if (json.toolData.projectiles && json.toolData.attack != 0 && !material.hasStats(MaterialTypes.PROJECTILE)) {
 		    TinkerRegistry.addMaterialStats(material,
 			    	new ProjectileMaterialStats()
 			    	);
@@ -293,8 +253,11 @@ public class MaterialData {
 	    
 	}
 	
-	private void makeFluid() {
-	    	    	    
+	public void makeFluid() {
+		if(!json.getMake("fluid")) {
+			return;	    			
+		}
+
 	    	String name = json.name.toLowerCase();
 		fluid = new FluidMolten(name, json.getFluidColorInt()); //, InfiniTiC.ICON_StillFluid, InfiniTiC.ICON_FlowingFluid);
 	    fluid.setUnlocalizedName(InfiniTiC.MODID + "." + name);  //For localization
@@ -312,15 +275,19 @@ public class MaterialData {
 		ResourceLocation regName = new ResourceLocation(InfiniTiC.MODID, name);
 	    block = new BlockMolten(fluid)
 		    .setUnlocalizedName(InfiniTiC.MODID + "." + name)  //For localization
-		    .setRegistryName(regName);
+		    .setRegistryName(regName)
+		    .setCreativeTab(null);
 	    ForgeRegistries.BLOCKS.register(block);
-	    ForgeRegistries.ITEMS.register(
-			new ItemBlock(block)
-				.setRegistryName(regName)
-		);
+	    fluidItem = new ItemBlock(block)
+				.setRegistryName(regName);
+	    ForgeRegistries.ITEMS.register(fluidItem);
 
 		FluidRegistry.addBucketForFluid(fluid);
 		
+	}
+
+	public void doIntegration() {
+	    if(integration != null) integration.integrate();	    	
 	}
 	
 }
