+++ src/main/java/lakmoore/infinitic/lib/data/MaterialJSON.java
@@ -1,12 +1,12 @@
-package me.johz.infinitic.lib.data;
+package lakmoore.infinitic.lib.data;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import me.johz.infinitic.InfiniTiC;
-import me.johz.infinitic.lib.errors.JSONValidationException;
-import me.johz.infinitic.lib.helpers.GenericHelper;
-import me.johz.infinitic.lib.helpers.NameConversionHelper;
+import lakmoore.infinitic.InfiniTiC;
+import lakmoore.infinitic.lib.errors.JSONValidationException;
+import lakmoore.infinitic.lib.helpers.GenericHelper;
+import lakmoore.infinitic.lib.helpers.NameConversionHelper;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.oredict.OreDictionary;
