+++ src/main/java/lakmoore/infinitic/lib/data/ToolDataJSON.java
@@ -1,8 +1,8 @@
-package me.johz.infinitic.lib.data;
+package lakmoore.infinitic.lib.data;
 
-import me.johz.infinitic.InfiniTiC;
-import me.johz.infinitic.lib.errors.JSONValidationException;
-import me.johz.infinitic.lib.helpers.GenericHelper;
+import lakmoore.infinitic.InfiniTiC;
+import lakmoore.infinitic.lib.errors.JSONValidationException;
+import lakmoore.infinitic.lib.helpers.GenericHelper;
 
 public class ToolDataJSON implements IJson {
