+++ src/main/java/lakmoore/infinitic/lib/helpers/ClipboardHelper.java
@@ -1,4 +1,4 @@
-package me.johz.infinitic.lib.helpers;
+package lakmoore.infinitic.lib.helpers;
 
 import java.awt.Toolkit;
 import java.awt.datatransfer.ClipboardOwner;
